refactor(tenants): consolidate pipeline step files, use narrow tenantAppProvider interface
Replaces the 5 individual pipeline_step_tenant_*.go files with a single
consolidated pipeline_step_tenant.go. Introduces tenantAppProvider interface
(GetService only) so unit tests can inject a lightweight fake instead of
requiring a full modular.Application. Adds *Std() wrapper factories for
step registry registration.
